Cree los arcivos iniciarJuego.js, selectorDePersonajes.js y Personajes.js al de iniciar... y selector... les puse funciones que estaba en el inde.js y quite los objetos class Pesonajes y Personaje... de Menu.js que no tenían que estar ahí y los puese en el archivo Personajes.js
