+++ codingFundamentalsJS/primerProyectoJS/index.js
@@ -1,113 +1,8 @@
- class Personaje {
-    constructor(nombre,especie,clase,almas,nivel,vida,aguilidad,ataque){
-        this.nom = nombre;
-        this.esp = especie;
-        this.clas = clase;
-        this.alm = almas;
-        this.niv = nivel;
-        this.heal = vida;
-        this.aguil = aguilidad;
-        this.daño = ataque;
-        //this.hollow = locura
-    }
-};
-
-
-
-let Personaje_1 = new Personaje("Ander","humano","Explorador",0,1,100,30,10)
-
-let Personaje_2 = new Personaje("Zasheida","humano","Explorador",0,1,100,30,)
-
-let Personaje_3 = new Personaje("Bryseis","Tiflin","Hechicero",0,1,100,20,15)
-
-let Personaje_4 = new Personaje("Damakos","Hechicero","",0,1,100,20,15)
-
-let Oponente1 = new Personaje("asdasd","asdasd","asdasd",0,1,90,10,7)
-
-const personaje_6 = new Personaje("","","",0,1)
-
-const personaje_7 = new Personaje("","","",0,1)
-
-const personaje_8 = new Personaje("","","",0,1)
-
-const personaje_9 = new Personaje("","","",0,1)
-
-function SelectorDePersonaje(per) {
-    switch (per) {
-        case Personaje_1.nom://podria hacer de ver como puedo juntar todos los caso pero mostrando el repectivo objeto que coresponda capaz cuntando todos los objetos en un arrays pero para mi que daria un problema muy gordo
-            console.log("Usted a seleccionado a");
-            console.table(Personaje_1)
-            let confir = prompt("¿está seguro de su elección? [si/no]");
-            validacionDePersonaje(confir);
-            jugador = Personaje_1;
-            break;
-        case Personaje_2.nom:
-            console.log("Usted a seleccionado a ",Personaje_2,"¿está seguro de su elección? [si/no]");
-            confir = prompt("");
-            validacionDePersonaje(confir);
-            jugador = Personaje_2;
-            break;
-        case Personaje_3.nom:
-            console.log("Usted a seleccionado a ",Personaje_3,"¿está seguro de su elección? [si/no]")
-            confir = prompt("")
-            validacionDePersonaje(confir)
-            jugador = Personaje_3;
-            break;
-        case Personaje_4:
-            console.log("Usted a seleccionado a ",Personaje_4,"¿está seguro de su elección? [si/no]")
-            confir = prompt("")
-            validacionDePersonaje(confir);
-            jugador = Personaje_4;
-            break;
-        default:
-            console.log("Su selección no es válida")
-            break;
-    }
-};
-
-function validacionDePersonaje(respuesta) {
-    if (respuesta == "si") {
-       bucleSelectorDePersonaje = false;
-       return console.log("Personaje seleccionado") 
-
-    }
-    else if (respuesta == "no") {     
-        return console.log("Personaje no seleccionado")
-
-    } else {
-        return console.log("No a papretado la tecla corecta") //no se como hacer para que se repetir que se vuelva a hce la pregunta de si y no
-    }
-};
+import { Personaje_1, Personaje_2, Personaje_3, Personaje_4, Oponente1 } from "./modelo/Menus";
+import {iniciarJuego} from ".funciones/iniciarJuego"
 
 // print("hola")
 
-function iniciarJuego(){
-    while (bucleDelMenu) {
-        console.log("Elija una opción, Comenzar partida, Opciones, Salir")
-
-        let menu = prompt("");//tambien podria poner una funcion o una variable con el saludo
-
-        if (menu == "Comenzar partida"){
-            bucleDelMenu = false/* prodria hacer que el usuario pueda volver para atras osea que while vuelva a ser true */
-            while (bucleSelectorDePersonaje) {
-
-                console.log("Seleccione un Personaje ",Personaje_1.nom,Personaje_2.nom,Personaje_3.nom,Personaje_4.nom)
-
-                let personajeSelecionado = prompt("")
-                SelectorDePersonaje(personajeSelecionado)
-                
-            }
-            
-        } else if (menu = "Opciones") {
-            console.log("Sección en desarrollo")
-        }
-        else{
-            console.log("No a selecionado ninguna opcion")
-        };
-        
-    };
-
-};
 prompt("Presione Enter");
 
 iniciarJuego();
